retrieveAndMerge: adopt a more reliable backfilling strategy
retrieveAndMerge was using *Whisper.FetchByAggregation method to retrieve data for merging and backfilling.
However, on production, we might have issues to retrieve data from the correct archive due to timing issues.
It was originally using a 60s buffer, that is not enought.
By switching to *Whisper.fetchFromArchive, it is guaranteed to retrieve data from desired archive based
on resolutions.